Refactor request components and types; migrate to new structure
