+++ src/components/request/StepNonBridalEvents.tsx
@@ -1,18 +1,19 @@
 "use client";
 import React, { useCallback } from "react";
-import type { BridalEventData } from "../types";
-import BridalEventCard from "./BridalEventCard";
-import { BridalEventField } from "../types";
+import type { EventData } from "@/types/request-types";
+import EventCard from "./EventCard";
+import { NonServiceField } from "@/types/request-types";
 
 interface Props {
-  events: BridalEventData[];
-  setEvents: React.Dispatch<React.SetStateAction<BridalEventData[]>>;
+  events: EventData[];
+  // important: make this a real React state setter
+  setEvents: React.Dispatch<React.SetStateAction<EventData[]>>;
   today: string;
   onBack: () => void;
   onNext: () => void;
 }
 
-export default function StepBridalEvents({
+export default function StepNonBridalEvents({
   events,
   setEvents,
   today,
@@ -20,10 +21,27 @@ export default function StepBridalEvents({
   onNext,
 }: Props) {
   const handleEventChange = useCallback(
-    (index: number, field: BridalEventField, value: string) => {
+    (index: number, field: NonServiceField, value: string) => {
       setEvents((prev) => {
         const next = [...prev];
         next[index] = { ...next[index], [field]: value };
+        return next;
+      });
+    },
+    [setEvents]
+  );
+
+  const handleServiceToggle = useCallback(
+    (index: number, service: string) => {
+      setEvents((prev) => {
+        const next = [...prev];
+        const svcs = next[index].services;
+        next[index] = {
+          ...next[index],
+          services: svcs.includes(service)
+            ? svcs.filter((s) => s !== service)
+            : [...svcs, service],
+        };
         return next;
       });
     },
@@ -36,26 +54,15 @@ export default function StepBridalEvents({
         Tell us about each event
       </h2>
 
-      <div className="border border-gray rounded-xl p-6 space-y-4">
-        <div className="flex items-center space-x-2 text-primary">
-          <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
-            <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-          </svg>
-          <span className="font-medium">Services Included</span>
-        </div>
-        <p className="text-sm text-dark mt-2">
-          Makeup and Hair services are automatically included
-        </p>
-      </div>
-
       <div className="space-y-8">
         {events.map((ev, idx) => (
-          <BridalEventCard
+          <EventCard
             key={ev.id}
             index={idx}
             event={ev}
             today={today}
             onChange={handleEventChange}
+            onToggleService={handleServiceToggle}
           />
         ))}
       </div>
